+++ crates/filter-list-manager/src/filters/parser/filter_contents_provider/io_provider.rs
@@ -1,5 +1,5 @@
 use super::FilterContentsProvider;
-use crate::io::fetch_by_schemes::fetch_by_scheme_with_content_check;
+use crate::io::fetch_by_schemes::{fetch_filter_by_scheme_with_content_check, FilterFetchPolicy};
 use crate::io::get_scheme;
 use crate::io::http::blocking_client::BlockingClient;
 use crate::FilterParserError;
@@ -21,7 +21,12 @@ impl FilterContentsProvider for IOProvider<'_> {
     fn get_filter_contents(&self, root_filter_url: &str) -> Result<String, FilterParserError> {
         let scheme = get_scheme(root_filter_url).unwrap_or_default();
 
-        fetch_by_scheme_with_content_check(root_filter_url, scheme.into(), self.get_http_client())
+        fetch_filter_by_scheme_with_content_check(
+            root_filter_url,
+            scheme.into(),
+            self.get_http_client(),
+            FilterFetchPolicy::RegularFilter,
+        )
     }
 
     fn get_http_client(&self) -> &BlockingClient {