+++ crates/filter-list-manager/src/io/http/blocking_client.rs
@@ -43,33 +43,24 @@ impl BlockingClient {
         Ok(Self { inner: client })
     }
 
-    /// Gets filter with special rules processing:
-    /// - StatusCode == 200, see below
-    ///
-    /// # Failure
-    ///
-    /// If status_code != 200, (e.g. 204), requests will fail with [`HttpClientError::Strict200Response`]
-    pub(crate) fn get_filter_bytes(&self, url: &str) -> Result<Bytes, HttpClientError> {
+    /// Gets filter bytes and status code
+    pub(crate) fn get_filter_bytes(
+        &self,
+        url: &str,
+    ) -> Result<(Bytes, StatusCode), HttpClientError> {
         let response = self
             .inner
             .get(url)
             .send()
-            .map_err(HttpClientError::make_network)?
-            .error_for_status()
             .map_err(HttpClientError::make_network)?;
 
         let status = response.status();
 
-        if status != StatusCode::OK {
-            return Err(HttpClientError::make_only_200_strict(
-                status,
-                url.to_owned(),
-            ));
-        }
-
-        response
+        let bytes = response
             .bytes()
-            .map_err(HttpClientError::make_body_recovery)
+            .map_err(HttpClientError::make_body_recovery)?;
+
+        Ok((bytes, status))
     }
 
     /// Gets a json from `url` and constructs type `T`